requirements.txt
asgiref==3.2.3
Django==3.0.2
django-nested-admin==3.2.4
python-dotenv==0.10.3
python-monkey-business==1.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
